Keep-alive time for A2DP source transport
This commit adds support for keeping A2DP transport alive for a given
number of seconds (or infinite time for negative value). One might use
this option for keeping external BT device in the powered state, or for
eliminating transport acquisition time lag when one is streaming short
audios in a repeatable matter.

Resolves #171